refactor: Reorganize StreamInf and ExtInf data structs
