+++ internal/node.go
@@ -20,6 +20,18 @@ type Node struct {
 	Prev, Next *Node
 }
 
+// The HLSElement data type holds the following attributes:
+//   - Name: The name of the Element (e.g. tag name).
+//   - URI: The Uniform Resource Identifier of the Element (if applicable).
+//   - Attrs: In-manifest Element attributes, in key-value format.
+//   - Details: Not-in-manifest Element attributes, in key-value format.
+type HLSElement struct {
+	Name    string
+	URI     string
+	Attrs   map[string]string
+	Details map[string]string
+}
+
 func (l *DoublyLinkedList) Insert(node *Node) {
 	if l.Head == nil {
 		l.Head = node
